refactor(compiler): move buildSystemPrompt into GenericStrategy
buildSystemPrompt was only used by GenericStrategy. Inlined the logic
into GenericStrategy.buildPrompt() and removed the export from
prompt-builder.ts. prompt-builder.ts now only contains buildScopedPrompt
(used by the transform stage).

Updated prompt-builder tests to test GenericStrategy.buildPrompt()
directly instead of the removed function.
