+++ modelguide-api/src/features/compiler/core/prompt-builder.ts
@@ -1,41 +1,12 @@
 /**
- * Prompt builder — assembles system prompt and per-step scoped prompts.
+ * Prompt builder — per-step scoped prompts.
  *
- * All content comes from the compiler input. The builder owns only the
- * section structure (ordering, markdown formatting, priority grouping).
- * Template is hardcoded for Phase 1.
- *
- * See PRD Appendix A.3.1 for concrete output examples.
+ * buildScopedPrompt is used by the transform stage to enrich each
+ * SOP step with matched guardrails.
  */
 
 import type { SopStep } from "@features/sops/sops.types";
-import type {
-  AgentSopInfo,
-  GuardrailPriority,
-  ParsedGuardrail,
-  ResolvedTool,
-} from "./types";
-
-/** Minimal SOP shape for prompt building (avoids Zod inference type conflicts). */
-interface SopForPrompt {
-  name: string;
-  description: string | null;
-  definition: {
-    steps: Array<{
-      order: number;
-      instruction: string;
-      required: boolean;
-      tool?: { resolvedName?: string };
-    }>;
-    metadata: {
-      escalationTriggers?: string[];
-    };
-  };
-}
-
-// ============================================================================
-// System prompt
-// ============================================================================
+import type { GuardrailPriority, ParsedGuardrail } from "./types";
 
 /** Priority display order (critical first). */
 const PRIORITY_ORDER: GuardrailPriority[] = [
@@ -44,98 +15,6 @@ const PRIORITY_ORDER: GuardrailPriority[] = [
   "medium",
   "low",
 ];
-
-/** Capitalize first letter of a priority for section headers. */
-function capitalize(s: string): string {
-  return s.charAt(0).toUpperCase() + s.slice(1);
-}
-
-/**
- * Build the system prompt — set once as the agent's `instructions`.
- *
- * Structure:
- * 1. agentConfig.description (role context)
- * 2. ## Workflow: {name} + description
- * 3. ## Tools — bullet list of available tool names
- * 4. ## Guardrails grouped by priority (Critical, High, Medium, Low)
- * 5. ## Escalation Triggers as bullet list
- */
-export function buildSystemPrompt(
-  agentDescription: string,
-  sop: SopForPrompt,
-  guardrails: ParsedGuardrail[],
-  tools: ResolvedTool[],
-  agentSops?: AgentSopInfo[],
-): string {
-  const sections: string[] = [];
-
-  // 1. Agent description (preamble)
-  sections.push(agentDescription);
-
-  // 1.5. Intent Classification (Step 0) — before workflow
-  const classificationSection = buildIntentClassificationSection(agentSops);
-  if (classificationSection) {
-    sections.push(classificationSection);
-  }
-
-  // 2. SOP context
-  sections.push(`## Workflow: ${sop.name}`);
-  if (sop.description) {
-    sections.push(sop.description);
-  }
-
-  // 2b. Steps
-  const steps = sop.definition.steps;
-  if (steps.length > 0) {
-    const stepLines = steps
-      .slice()
-      .sort((a, b) => a.order - b.order)
-      .map((s) => {
-        const toolSuffix = s.tool?.resolvedName
-          ? ` → \`${s.tool.resolvedName}\``
-          : "";
-        const requiredTag = s.required ? " *(required)*" : "";
-        return `${s.order}. ${s.instruction}${toolSuffix}${requiredTag}`;
-      });
-    const enforcement =
-      "Follow these steps in order. Every step must be reflected in your response — do not skip steps.";
-    sections.push(`### Steps\n${enforcement}\n${stepLines.join("\n")}`);
-  }
-
-  // 3. Tools
-  if (tools.length > 0) {
-    sections.push("## Tools");
-    sections.push(tools.map((t) => `- ${t.resolvedName}`).join("\n"));
-  }
-
-  // 4. Guardrails grouped by priority
-  const guardrailsByPriority = groupByPriority(guardrails);
-  const guardrailSections: string[] = [];
-
-  for (const priority of PRIORITY_ORDER) {
-    const group = guardrailsByPriority.get(priority);
-    if (!group || group.length === 0) continue;
-
-    guardrailSections.push(`### ${capitalize(priority)}`);
-    for (const g of group) {
-      guardrailSections.push(`**${g.name}:** ${g.content}`);
-    }
-  }
-
-  if (guardrailSections.length > 0) {
-    sections.push("## Guardrails");
-    sections.push(guardrailSections.join("\n\n"));
-  }
-
-  // 5. Escalation triggers
-  const triggers = sop.definition.metadata.escalationTriggers;
-  if (triggers && triggers.length > 0) {
-    sections.push("## Escalation Triggers");
-    sections.push(triggers.map((t) => `- ${t}`).join("\n"));
-  }
-
-  return sections.join("\n\n");
-}
 
 // ============================================================================
 // Scoped prompt (per-step)
@@ -178,56 +57,4 @@ export function buildScopedPrompt(
   }
 
   return sections.join("\n\n");
-}
-
-// ============================================================================
-// Intent classification
-// ============================================================================
-
-/**
- * Build the "## Intent Classification (Step 0)" section.
- * Returns null if no agent SOPs are available.
- */
-export function buildIntentClassificationSection(
-  agentSops?: AgentSopInfo[],
-): string | null {
-  if (!agentSops || agentSops.length === 0) return null;
-
-  const sopList = agentSops
-    .map((sop) => {
-      const desc = sop.description ? `: ${sop.description}` : "";
-      return `- \`${sop.slug}\` — ${sop.name}${desc}`;
-    })
-    .join("\n");
-
-  return `## Intent Classification (Step 0)
-
-Before executing any workflow steps, classify the customer's intent by calling \`core_classify_sop\`.
-
-Available SOPs:
-${sopList}
-
-Call \`core_classify_sop\` with:
-- \`sop_slug\`: the matching SOP slug from the list above (or null if no match)
-- \`confidence\`: your confidence in the classification (0.0–1.0)
-- \`session_id\`: the current session ID
-
-Do not wait for the response before proceeding with the conversation.`;
-}
-
-// ============================================================================
-// Helpers
-// ============================================================================
-
-function groupByPriority(
-  guardrails: ParsedGuardrail[],
-): Map<GuardrailPriority, ParsedGuardrail[]> {
-  const map = new Map<GuardrailPriority, ParsedGuardrail[]>();
-  for (const g of guardrails) {
-    const priority = g.config.priority;
-    const group = map.get(priority) ?? [];
-    group.push(g);
-    map.set(priority, group);
-  }
-  return map;
 }
